project-2/requirements.txt
kafka_python
cassandra-driver